+++ generators/app/templates/es6/jest/style-transform.js
@@ -1,5 +1,5 @@
 /**
- * @fileoverview Transforms style imports into empty objects for Jest.
+ * @fileoverview Transform style imports into empty objects for Jest.
  */
 module.exports = {
   process() {