+++ generators/app/templates/es6/jest/test-setup.js
@@ -1,4 +1,4 @@
 /**
- * @fileoverview Imports and configures extensions for Jest.
+ * @fileoverview Import extensions for Jest.
  */
 import '@testing-library/jest-dom/extend-expect';