+++ generators/app/templates/es6/src/polyfills.js
@@ -1,5 +1,5 @@
 /**
- * @fileoverview Imports polyfills for Webpack config entry point.
+ * @fileoverview Import polyfills for Webpack config entry point.
  */
 import 'core-js/stable';
 import 'regenerator-runtime/runtime';