+++ generators/app/templates/typescript/jest/jest-dom.d.ts
@@ -1,5 +1,5 @@
 /**
- * Imports jest-dom matcher extensions.
+ * Import jest-dom matcher extensions.
  * @packageDocumentation
  */
 import '@testing-library/jest-dom/extend-expect';