+++ generators/app/templates/typescript/jest/test-setup.ts
@@ -1,5 +1,5 @@
 /**
- * Imports and configures extensions for Jest.
+ * Import extensions for Jest.
  * @packageDocumentation
  */
 import '@testing-library/jest-dom/extend-expect';