+++ generators/app/templates/typescript/src/components/app/App.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import { hot } from 'react-hot-loader/root';
+import Home from '@src/components/home';
 import Missing from '@src/components/missing';
-import routes from '@src/routes/app-routes';
 import Header from './header';
 import Footer from './footer';
 
@@ -10,9 +10,7 @@ const App: React.FC = () => (
   <>
     <Header />
     <Switch>
-      {routes.map((route) => (
-        <Route exact key={route.path} {...route} />
-      ))}
+      <Route exact path="/" component={Home} />
       <Route component={Missing} />
     </Switch>
     <Footer />